Remove the confusing `--verbose` vs. `--debug` CLI options and log more.
This sets the default loglevel to INFO (was CRITICAL) and reworks all of
the sal-submit logging to make more stuff INFO. It then removes the
`--verbose` arg.

This was originally meant to not spew a bunch of Sal stuff at the end of
Munki runs, but we've subsequently decoupled Sal submissions from Munki,
so that's not really a goal any longer.